project-rag-brain/enterprise-brain/requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
langchain-experimental
langgraph
langgraph-checkpoint-sqlite
python-dotenv
chromadb
sentence-transformers
streamlit
duckduckgo-search
uvicorn
pydantic-settings
pypdf
fastapi
python-multipart
rank_bm25
psutil
markitdown